improve accuracy of test coverage
