+++ deno.jsonc
@@ -22,7 +22,7 @@
   },
   "tasks": {
     "test": "deno fmt --check && deno lint && deno bundle src/mod.ts && deno test --allow-read --parallel src/ tests/",
-    "coverage": "rm -rf .coverage && deno test --reporter=dot --parallel --allow-read --coverage=.coverage src/ && deno coverage --exclude=fixtures --exclude=_test --lcov --output=lcov.info .coverage && deno run --allow-read https://deno.land/x/code_coverage@0.2.0/cli.ts"
+    "coverage": "rm -rf .coverage && deno test --reporter=dot --parallel --allow-read --coverage=.coverage src/ && deno coverage --exclude=fixtures --exclude=_test --lcov --output=lcov.info .coverage && deno run --allow-read https://deno.land/x/code_coverage@0.2.2/cli.ts"
   },
   "lock": false
 }
