+++ src/dev_deps.ts
@@ -10,6 +10,7 @@ export {
   fail,
 } from "https://deno.land/std@0.152.0/testing/asserts.ts";
 export * as mock from "https://deno.land/std@0.152.0/testing/mock.ts";
+export { assertSnapshot } from "https://deno.land/std@0.152.0/testing/snapshot.ts";
 
 export { assert } from "https://deno.land/x/conditional_type_checks@1.0.6/mod.ts";
 export type {
