+++ src/schema/slack/functions/_scripts/src/write_function_files.ts
@@ -9,7 +9,7 @@ import { FunctionRecord } from "./types.ts";
 const VALID_FILENAME_REGEX = /^[0-9a-zA-Z_\-]+$/;
 
 async function main() {
-  const slackFunctions: FunctionRecord[] = await getSlackFunctions();
+  const slackFunctions: FunctionRecord[] = await _internals.getSlackFunctions();
 
   // Sorting alphabetically cause only a monster would generate these in a random order
   slackFunctions.sort((a, b) => a.callback_id.localeCompare(b.callback_id));
@@ -32,11 +32,13 @@ async function main() {
       const filename = `../${functionRecord.callback_id}.ts`;
       const testFilename = `../${functionRecord.callback_id}_test.ts`;
 
-      const templateString = SlackFunctionTemplate(functionRecord);
-      const templateTestString = SlackTestFunctionTemplate(functionRecord);
+      const templateString = _internals.SlackFunctionTemplate(functionRecord);
+      const templateTestString = _internals.SlackTestFunctionTemplate(
+        functionRecord,
+      );
 
-      await Deno.writeTextFile(filename, templateString);
-      await Deno.writeTextFile(testFilename, templateTestString);
+      await _internals.writeTextFile(filename, templateString);
+      await _internals.writeTextFile(testFilename, templateTestString);
     }),
   );
 
@@ -44,12 +46,27 @@ async function main() {
     `Generated ${slackFunctions.length} Slack functions with their unit tests`,
   );
 
-  const modString = SlackFunctionModTemplate(slackFunctions);
+  const modString = _internals.SlackFunctionModTemplate(slackFunctions);
 
-  await Deno.writeTextFile("../mod.ts", modString);
+  await _internals.writeTextFile("../mod.ts", modString);
   console.log("Updated functions module export");
 }
 
+export const _internals = {
+  main,
+  getSlackFunctions,
+  SlackFunctionModTemplate,
+  SlackFunctionTemplate,
+  SlackTestFunctionTemplate,
+  writeTextFile: (
+    path: string,
+    data: string,
+    options?: Deno.WriteFileOptions,
+  ): ReturnType<typeof Deno.writeTextFile> => {
+    return Deno.writeTextFile(path, data, options);
+  },
+};
+
 if (import.meta.main) {
-  main();
+  _internals.main();
 }
