+++ src/schema/slack/functions/_scripts/src/write_function_files_test.ts
@@ -1,0 +1,6 @@
+import { assertExists } from "../../../../../dev_deps.ts";
+import * as writeFunctionFiles from "./write_function_files.ts";
+
+Deno.test("Include all content of write_function_files.ts in code coverage", () => {
+  assertExists(writeFunctionFiles);
+});
